skills/paper-translator/requirements.txt
# PDF解析
PyMuPDF>=1.24.0
pdfplumber>=0.11.0

# LLM调用
anthropic>=0.40.0
openai>=1.50.0

# 图像处理
Pillow>=10.0.0

# 公众号发布
requests>=2.31.0

# Markdown处理
markdown>=3.5.0
beautifulsoup4>=4.12.0

# 配置管理
pyyaml>=6.0.0
python-dotenv>=1.0.0

# 工具库
httpx>=0.27.0
tenacity>=8.2.0
